+++ ios/Classes/yolox.cpp
@@ -401,6 +401,8 @@ static void draw_objects(const cv::Mat& bgr, const std::vector<Object>& objects)
 
 extern "C" __attribute__((visibility("default"))) __attribute__((used)) void initYolox(char *modelPath, char *paramPath)
 {
+    yolox.clear();
+
     // Focus in yolov5
     yolox.register_custom_layer("YoloV5Focus", YoloV5Focus_layer_creator);