+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.1
 
 - The Width of the stream's CameraImage is width from the length of the pixels. #45
+- improvement detect method #46
 
 ## 0.3.0